+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Previous Version: 0.0.0b3
 
 ## New Features
 
+- expose the SAScore threshold as a parameter of the exploration tree (`tree.params.sascoreMax`).
+
 ## Changes